pyproject.toml
[project]
name = "immigration-ai-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "faiss-cpu>=1.13.2",
    "fastapi>=0.135.1",
    "groq>=1.1.1",
    "langchain>=1.2.13",
    "langchain-community>=0.4.1",
    "langchain-huggingface>=1.2.1",
    "langchain-text-splitters>=1.1.1",
    "openai>=2.29.0",
    "pyairtable>=3.3.0",
    "pypdf>=6.9.1",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "resend>=2.25.0",
    "scikit-learn>=1.7.2",
    "streamlit>=1.55.0",
    "uvicorn>=0.42.0",
]
